all_combos() for n-consistency
Added the all_combos function to help find n-consistent assignments for a CSP problem.

This may not be a good choice. Instead of trying to solve certain CSP/COP problems with inference, it may be a better idea to jump straight to a backtracking algo.

This algo isn't super efficient, but it can be helpful in debugging.